refactor: QuestionEvaluations(질문별 평가점수와 순위를 저장)클래스를 도메인 계층으로 변경
